requirements.txt
fastapi
uvicorn
streamlit
transformers
diffusers
torch
sqlalchemy
psycopg2-binary
Pillow
requests
